test multiarch build

+++ create_image.sh
@@ -1,6 +1,7 @@
 #!/bin/bash -e
 
 minimal_kmsdrm=${KMSDRM:-0} #do not install x11 or nvidia driver
+arch=${BUILDARCH:-amd64}
 onlybuild=0 #use existing workdir and only build image
 
 distro="debian" #devuan or debian
@@ -35,10 +36,16 @@ fi
 #for releases - without a target dir this means we're doing the build on github
 targetdir="/mnt/nas-quake/quake_bootable"
 if [ -d "$targetdir" ];then
-	imagename="${imagebase}${imagesuffix}-$(date +"%Y-%m-%d").img"
-	imagelatestname="${imagebase}${imagesuffix}-latest"
+	imagename="${imagebase}${imagesuffix}-${arch}-$(date +"%Y-%m-%d").img"
+	imagelatestname="${imagebase}${imagesuffix}-${arch}-latest"
 else
-	imagename="${imagebase}${imagesuffix}.img"
+	imagename="${imagebase}${imagesuffix}-${arch}.img"
+fi
+
+if [ "$arch" == "i386" ];then
+	cpuarch="686"
+else
+	cpuarch="$arch"
 fi
 
 lvmdir="$currentdir/lvm"
@@ -60,7 +67,7 @@ tintrc="$currentdir/resources/tint2rc"
 modprobe="$currentdir/resources/modprobe.d"
 issueappend="$currentdir/resources/issue.append"
 
-packages="os-prober util-linux iputils-ping openssh-client file git sudo build-essential libgl1-mesa-dri libpcre3-dev terminfo procps vim-tiny unzip zstd alsa-utils grub2 cpufrequtils fbset chrony cloud-utils parted lvm2 gdisk initramfs-tools fdisk intel-microcode amd64-microcode firmware-linux firmware-linux-nonfree firmware-linux-free libarchive-tools linux-image-amd64 ntfs-3g nfs-common "
+packages="os-prober util-linux iputils-ping openssh-client file git sudo build-essential libgl1-mesa-dri libpcre3-dev terminfo procps vim-tiny unzip zstd alsa-utils grub2 cpufrequtils fbset chrony cloud-utils parted lvm2 gdisk initramfs-tools fdisk intel-microcode amd64-microcode firmware-linux firmware-linux-nonfree firmware-linux-free libarchive-tools linux-image-${cpuarch} ntfs-3g nfs-common "
 packages_nox11="ifupdown dhcpcd-base"
 packages_x11=" xserver-xorg-legacy xserver-xorg-core xserver-xorg-video-amdgpu xserver-xorg-input-all xinit iw connman connman-gtk feh xterm obconf openbox tint2 fbautostart menu python3-xdg xdg-utils lxrandr dex chromium pasystray pavucontrol pipewire pipewire-pulse wireplumber rtkit dex x11-xserver-utils dbus-x11 dbus-bin imagemagick pcmanfm gvfs-backends lxpolkit "
 
@@ -111,9 +118,9 @@ if [ $onlybuild -eq 0 ] || [ ! -d "$workdir/usr" ];then
 	mkdir -p "$workdir"
 	
 	if [ "$distro" = "devuan" ];then
-		sudo debootstrap --include="devuan-keyring gnupg wget ca-certificates" --exclude="debian-keyring" --no-check-gpg --variant=minbase $release "$workdir" http://dev.beard.ly/devuan/merged/
+		sudo debootstrap --arch=${arch} --include="devuan-keyring gnupg wget ca-certificates" --exclude="debian-keyring" --no-check-gpg --variant=minbase $release "$workdir" http://dev.beard.ly/devuan/merged/
 	else
-		sudo debootstrap --include="debian-keyring gnupg wget ca-certificates" --exclude="devuan-keyring" --no-check-gpg --variant=minbase $release "$workdir" https://deb.debian.org/debian/
+		sudo debootstrap --arch=${arch} --include="debian-keyring gnupg wget ca-certificates" --exclude="devuan-keyring" --no-check-gpg --variant=minbase $release "$workdir" https://deb.debian.org/debian/
 	fi
 	export distro
 	
@@ -133,7 +140,7 @@ if [ $onlybuild -eq 0 ] || [ ! -d "$workdir/usr" ];then
 	sudo cp -fR "$quakedir" "$workdir/quake"
 	sudo cp -f "$background" "$workdir/background.png"
 
-	sudo --preserve-env=nquakeresourcesurl,nquakeresourcesurl_backup,nquakezips,ezquakegitrepo,packages,distro,minimal_kmsdrm chroot "$workdir" bash -e -c '
+	sudo --preserve-env=nquakeresourcesurl,nquakeresourcesurl_backup,nquakezips,ezquakegitrepo,packages,distro,minimal_kmsdrm,arch,cpuarch chroot "$workdir" bash -e -c '
 	
 	#configure hostname
 	echo "127.0.1.1 '$mediahostname'" >> /etc/hosts
@@ -233,8 +240,10 @@ if [ $onlybuild -eq 0 ] || [ ! -d "$workdir/usr" ];then
 	echo "building ezquake"
 	. /home/quakeuser/.profile
 	#drop march to nehalem instead of native
-	export CFLAGS="${CFLAGS} -march=nehalem"
-	export LDFLAGS="${CFLAGS}"
+	if [ "$arch" == "amd64" ];then
+		export CFLAGS="${CFLAGS} -march=nehalem"
+		export LDFLAGS="${CFLAGS}"
+  fi
 	rm -rf /home/quakeuser/build
 	mkdir /home/quakeuser/build
 	git clone --depth=1 $ezquakegitrepo /home/quakeuser/build/ezquake-source-official
@@ -262,19 +271,25 @@ if [ $onlybuild -eq 0 ] || [ ! -d "$workdir/usr" ];then
 	fi
 
 	#openrazer and kernel headers	
-	apt-get -qy install openrazer-driver-dkms linux-headers-amd64
+	apt-get -qy install openrazer-driver-dkms linux-headers-${cpuarch}
 
-	if [ "$minimal_kmsdrm" != "1" ];then
-		#install afterquake
-		mkdir -p /home/quakeuser/quake-afterquake
-		wget -qO /tmp/aq.zip https://fte.triptohell.info/moodles/linux_amd64/afterquake.zip
-		unzip /tmp/aq.zip -d /home/quakeuser/quake-afterquake
-		rm /tmp/aq.zip
-		chown quakeuser:quakeuser -Rf /home/quakeuser/quake-afterquake
-
-		#install nvidia and openrazer drivers
-		apt-get -qy install nvidia-driver nvidia-settings
-	fi
+  if [ "$arch" == "i386" ] || [ "$arch" == "amd64" ];then
+   	if [ "$minimal_kmsdrm" != "1" ];then
+   		#install afterquake
+   		mkdir -p /home/quakeuser/quake-afterquake
+   		if [ "$arch" == "i386" ];then
+   			wget -qO /tmp/aq.zip https://fte.triptohell.info/moodles/linux_x86/afterquake.zip
+			else
+   			wget -qO /tmp/aq.zip https://fte.triptohell.info/moodles/linux_amd64/afterquake.zip
+   		fi
+   		unzip /tmp/aq.zip -d /home/quakeuser/quake-afterquake
+   		rm /tmp/aq.zip
+   		chown quakeuser:quakeuser -Rf /home/quakeuser/quake-afterquake
+   
+   		#install nvidia and openrazer drivers
+   		apt-get -qy install nvidia-driver nvidia-settings
+   	fi
+  fi
 
 	#update nquake resources
 	echo "updating nquake resources..."
@@ -331,7 +346,7 @@ if [ $onlybuild -eq 0 ] || [ ! -d "$workdir/usr" ];then
 	rm -rf /var/log/*
 
 	#let debootstick install this
-	apt -y purge linux-image-amd64 || true
+	apt -y purge linux-image-${cpuarch} || true
 	apt-get -qy autopurge || true
 	
 	#remove temporary resolv.conf
